Add files via upload
Here are the files associated with the Wrangle This project. It includes two csv files, one tsv file, 1 txt file, 2 pdfs and 1 ipynb file. Together these will show the how this project was completed and the steps associated with it. We used the twitter archive to gather and asses the weRatedogs data. After the assessment stage, we cleaned the data as needed and performed programmatic or manual queries to clean the data. Then we were able to visualize it and report on our findings.